requirements.txt
pymupdf==1.24.9
pillow==10.4.0
paddlepaddle==2.6.2
paddleocr==2.7.3
numpy==1.26.4
flask==3.0.3
gunicorn==22.0.0
